Extract `TestRandom` from `ItemEnhancement`
