+++ .Lib9c.Tests/Action/CombinationEquipmentTest.cs
@@ -36,7 +36,7 @@ namespace Lib9c.Tests.Action
                 )
             );
             _sheets = TableSheetsImporter.ImportSheets();
-            _random = new ItemEnhancementTest.TestRandom();
+            _random = new TestRandom();
             _tableSheets = new TableSheets(_sheets);
             var agentState = new AgentState(_agentAddress);
             agentState.avatarAddresses[0] = _avatarAddress;
@@ -181,7 +181,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = new State(),
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
         }
 
@@ -232,7 +232,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
         }
 
@@ -317,7 +317,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
         }
 
@@ -354,7 +354,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
         }
 
@@ -388,7 +388,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
         }
     }
