+++ .Lib9c.Tests/Action/HackAndSlash2Test.cs
@@ -96,7 +96,7 @@ namespace Lib9c.Tests.Action
                 .First(r => r.ItemSubType == ItemSubType.FullCostume)
                 .Id;
             var costume =
-                ItemFactory.CreateItem(_tableSheets.ItemSheet[costumeId], new ItemEnhancementTest.TestRandom());
+                ItemFactory.CreateItem(_tableSheets.ItemSheet[costumeId], new TestRandom());
             previousAvatarState.inventory.AddItem(costume);
 
             var state = _initialState.SetState(_avatarAddress, previousAvatarState.Serialize());
@@ -119,7 +119,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
                 Rehearsal = false,
             });
 
@@ -171,7 +171,7 @@ namespace Lib9c.Tests.Action
                 {
                     PreviousStates = _initialState,
                     Signer = _agentAddress,
-                    Random = new ItemEnhancementTest.TestRandom(),
+                    Random = new TestRandom(),
                     Rehearsal = false,
                 })
             );
@@ -199,7 +199,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = new State(),
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -228,7 +228,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -259,7 +259,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -291,7 +291,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -331,7 +331,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -362,7 +362,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -408,7 +408,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -440,7 +440,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -485,7 +485,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -531,7 +531,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -568,7 +568,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = state,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             }));
 
             Assert.Null(action.Result);
@@ -631,7 +631,7 @@ namespace Lib9c.Tests.Action
             {
                 PreviousStates = _initialState,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
                 Rehearsal = false,
             });
