+++ .Lib9c.Tests/Action/ItemEnhancement3Test.cs
@@ -29,7 +29,7 @@ namespace Lib9c.Tests.Action
         public ItemEnhancement3Test()
         {
             var sheets = TableSheetsImporter.ImportSheets();
-            _random = new ItemEnhancementTest.TestRandom();
+            _random = new TestRandom();
             _tableSheets = new TableSheets(sheets);
             var privateKey = new PrivateKey();
             _agentAddress = privateKey.PublicKey.ToAddress();
