+++ .Lib9c.Tests/Action/ItemEnhancementTest.cs
@@ -510,35 +510,5 @@ namespace Lib9c.Tests.Action
 
             Assert.Equal(updatedAddresses.ToImmutableHashSet(), nextState.UpdatedAddresses);
         }
-
-        public class TestRandom : IRandom
-        {
-            private readonly System.Random _random = new System.Random();
-
-            public int Next()
-            {
-                return _random.Next();
-            }
-
-            public int Next(int maxValue)
-            {
-                return _random.Next(maxValue);
-            }
-
-            public int Next(int minValue, int maxValue)
-            {
-                return _random.Next(minValue, maxValue);
-            }
-
-            public void NextBytes(byte[] buffer)
-            {
-                _random.NextBytes(buffer);
-            }
-
-            public double NextDouble()
-            {
-                return _random.NextDouble();
-            }
-        }
     }
 }
