+++ .Lib9c.Tests/Action/RedeemCodeTest.cs
@@ -92,7 +92,7 @@ namespace Lib9c.Tests.Action
                 PreviousStates = initialState,
                 Rehearsal = false,
                 Signer = _agentAddress,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
             });
 
             // Check target avatar & agent
