+++ .Lib9c.Tests/Action/Sell2Test.cs
@@ -120,7 +120,7 @@
                     PreviousStates = previousStates,
                     Rehearsal = false,
                     Signer = _agentAddress,
-                    Random = new ItemEnhancementTest.TestRandom(),
+                    Random = new TestRandom(),
                 });
 
                 productCount++;
