+++ .Lib9c.Tests/Action/Sell3Test.cs
@@ -131,7 +131,7 @@
                     PreviousStates = previousStates,
                     Rehearsal = false,
                     Signer = _agentAddress,
-                    Random = new ItemEnhancementTest.TestRandom(),
+                    Random = new TestRandom(),
                 });
 
                 productCount++;
