+++ .Lib9c.Tests/Action/SellCancellation2Test.cs
@@ -119,7 +119,7 @@ namespace Lib9c.Tests.Action
             {
                 BlockIndex = 0,
                 PreviousStates = _initialState,
-                Random = new ItemEnhancementTest.TestRandom(),
+                Random = new TestRandom(),
                 Rehearsal = false,
                 Signer = _agentAddress,
             });
