+++ .Lib9c.Tests/Model/StageSimulatorTest.cs
@@ -21,7 +21,7 @@ namespace Lib9c.Tests.Model
         public StageSimulatorTest()
         {
             _tableSheets = new TableSheets(TableSheetsImporter.ImportSheets());
-            _random = new ItemEnhancementTest.TestRandom();
+            _random = new TestRandom();
 
             _avatarState = new AvatarState(
                 default,
